requirements.txt
openai==1.13.3
streamlit==1.32.0
audio_recorder_streamlit
streamlit-float
python-dotenv
langchain==0.1.16
chromadb
tiktoken
pypdf==4.2.0
httpx==0.27.2
speechrecognition